+++ src/Am.Keyward.Infrastructure/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Am.Keyward.Infrastructure;
-
-public class Class1
-{
-
-}